@@ -1,4 +1,4 @@
-// swift-tools-version: 5.7
+// swift-tools-version: 6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -6,7 +6,7 @@ import PackageDescription
 let package = Package(
   name: "AsyncMultiplexImage",
   platforms: [
-    .iOS(.v15),
+    .iOS(.v16),
   ],
   products: [
     // Products define the executables and libraries a package produces, and make them visible to other packages.
@@ -46,5 +46,6 @@ let package = Package(
       name: "AsyncMultiplexImageTests",
       dependencies: ["AsyncMultiplexImage"]
     ),
-  ]
+  ],
+  swiftLanguageModes: [.v6]
 )
